dev-java/jfreesvg: Version bump 3.2 -> 3.3
New release less demos, thanks to author!
https://github.com/jfree/jfreesvg/issues/12

Package-Manager: Portage-2.3.13, Repoman-2.3.4

+++ dev-java/jfreesvg/jfreesvg-9999.ebuild
@@ -30,6 +30,8 @@ S="${WORKDIR}/${P}"
 
 java_prepare() {
 	# Drop demos per https://github.com/jfree/jfreesvg/issues/12
-	rm -r src/main/java/org/jfree/graphics2d/demo \
-		|| die "Failed to remove demos"
+	if [[ ${PV} == 3.2 ]]; then
+		rm -r src/main/java/org/jfree/graphics2d/demo \
+			|| die "Failed to remove demos"
+	fi
 }
